moved coach_team_info to folder
