+++ lib/views/mobile/screens/coach_view.dart
@@ -5,7 +5,7 @@ import "package:orbit_standard_library/orbit_standard_library.dart";
 import "package:scouting_frontend/models/team_model.dart";
 import "package:scouting_frontend/net/hasura_helper.dart";
 import "package:scouting_frontend/views/constants.dart";
-import "package:scouting_frontend/views/mobile/screens/coach_team_info_data.dart";
+import "package:scouting_frontend/views/mobile/screens/coach_team_info/coach_team_info_data.dart";
 import "package:scouting_frontend/views/mobile/side_nav_bar.dart";
 import "package:scouting_frontend/views/pc/compare/compare_screen.dart";
