+++ lib/views/pc/picklist/pick_list_widget.dart
@@ -1,7 +1,7 @@
 import "package:flutter/material.dart";
 import "package:scouting_frontend/models/team_data/all_team_data.dart";
 import "package:scouting_frontend/views/constants.dart";
-import "package:scouting_frontend/views/mobile/screens/coach_team_info_data.dart";
+import "package:scouting_frontend/views/mobile/screens/coach_team_info/coach_team_info_data.dart";
 import "package:scouting_frontend/views/pc/picklist/pick_list_screen.dart";
 import "package:scouting_frontend/views/pc/team_info/team_info_screen.dart";
 import "package:flutter_switch/flutter_switch.dart";
